SE PUEDE GENERAR CITAS, SE AÑADIO FUNCIONALIDADES AL USUARIO, AHORA PUEDE LISTAR, DESACTIVAR Y SE AGREGARON DOS CAMPOS ADICIONALES

+++ src/main/java/webvet/v1/infraestructure/repository/UsuarioRepository.java
@@ -5,6 +5,6 @@ import webvet.v1.infraestructure.entity.UsuarioEntity;
 
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
+public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Integer> {
     Optional<UsuarioEntity> findByUsername(String username);
 }
